+++ src/main/java/dooya/see/domain/member/exception/AuthenticateException.java
@@ -1,4 +1,4 @@
-package dooya.see.domain.member;
+package dooya.see.domain.member.exception;
 
 public class AuthenticateException extends RuntimeException {
     public AuthenticateException(String message) {
