+++ src/main/java/dooya/see/domain/member/exception/DuplicateProfileException.java
@@ -1,4 +1,4 @@
-package dooya.see.domain.member;
+package dooya.see.domain.member.exception;
 
 public class DuplicateProfileException extends RuntimeException {
     public DuplicateProfileException(String message) {
